+++ reprounzip-qt/reprounzip_qt/gui/unpack.py
@@ -5,12 +5,13 @@
 from __future__ import division, print_function, unicode_literals
 
 import os
+import subprocess
 
 from PyQt4 import QtCore, QtGui
 
 import reprounzip_qt.reprounzip_interface as reprounzip
 from reprounzip_qt.gui.common import ROOT, ResizableStack, \
-    handle_error, error_msg
+    handle_error, error_msg, parse_ports
 
 
 class UnpackerOptions(QtGui.QWidget):
@@ -79,6 +80,27 @@ class DockerOptions(UnpackerOptions):
         self.root.addItems(ROOT.TEXT)
         self.add_row("Elevate privileges:", self.root)
 
+        try:
+            cmd = ['docker-machine', 'ls', '-q']
+            query = subprocess.Popen(cmd, stdout=subprocess.PIPE)
+            out, _ = query.communicate()
+            if query.returncode != 0:
+                raise subprocess.CalledProcessError(query.returncode, cmd)
+            self.machine = QtGui.QComboBox(editable=False)
+            if 'DOCKER_HOST' in os.environ:
+                self.machine.addItem("Custom config from environment", None)
+            else:
+                self.machine.addItem("Default (no machine)", None)
+            for machine in out.splitlines():
+                machine = machine.strip()
+                if machine:
+                    self.machine.addItem(machine.decode('utf-8', 'replace'),
+                                         machine)
+        except (OSError, subprocess.CalledProcessError):
+            self.machine = QtGui.QComboBox(editable=False, enabled=False)
+            self.machine.addItem("docker-machine unavailable", None)
+        self.add_row("docker-machine:", self.machine)
+
         self.image = QtGui.QLineEdit(placeholderText='detect')
         self.add_row("Base image:", self.image)
 
@@ -91,6 +113,10 @@ class DockerOptions(UnpackerOptions):
 
     def options(self):
         options = super(DockerOptions, self).options()
+
+        if self.machine.currentIndex() != -1:
+            options['docker-machine'] = self.machine.itemData(
+                self.machine.currentIndex())
 
         options['root'] = ROOT.INDEX_TO_OPTION[self.root.currentIndex()]
 
@@ -124,6 +150,11 @@ class VagrantOptions(UnpackerOptions):
         self.gui = QtGui.QCheckBox("Enable local GUI")
         self.add_row("GUI:", self.gui)
 
+        self.ports = QtGui.QLineEdit(
+            '',
+            toolTip="Space-separated host:guest port mappings")
+        self.add_row("Expose ports:", self.ports)
+
         self.use_chroot = QtGui.QCheckBox("use chroot and prefer packed files "
                                           "over the virtual machines' files",
                                           checked=True)
@@ -148,6 +179,13 @@ class VagrantOptions(UnpackerOptions):
 
         if self.gui.isChecked():
             options['args'].append('--use-gui')
+
+        ports = parse_ports(self.ports.text(), self)
+        if ports is None:
+            return None
+        for host, container, proto in ports:
+            options['args'].append('--expose-port=%s:%s/%s' % (
+                                   host, container, proto))
 
         if not self.use_chroot.isChecked():
             options['args'].append('--dont-use-chroot')